Extract AI hooks and refactor AI view component
- Refactored ai.tsx with improved modular architecture and hook integration
- Created ai-backup.tsx for safe rollback during refactoring process
- Extracted use-ai-generation.ts hook for AI generation logic and state management
- Added use-ai-history.ts hook for AI generation history and tracking
- Improved separation of concerns with dedicated hooks for AI functionality
- Enhanced maintainability and reusability of AI component logic
- Better type safety and error handling in AI generation workflow
