+++ qcut/apps/web/src/components/editor/media-panel/views/ai.tsx
@@ -13,7 +13,7 @@ import {
   X,
   Check,
 } from "lucide-react";
-import { useState, useEffect, useRef, Fragment } from "react";
+import { useState, useRef } from "react";
 import { Textarea } from "@/components/ui/textarea";
 import {
   Select,
@@ -25,147 +25,71 @@ import {
 import { Label } from "@/components/ui/label";
 import { Button } from "@/components/ui/button";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import {
-  generateVideo,
-  generateVideoFromImage,
-  handleApiError,
-  getGenerationStatus,
-  ProgressCallback,
-} from "@/lib/ai-video-client";
-import { AIVideoOutputManager } from "@/lib/ai-video-output";
+
 import { useTimelineStore } from "@/stores/timeline-store";
-import { useAsyncMediaStoreActions } from "@/hooks/use-async-media-store";
 import { useProjectStore } from "@/stores/project-store";
 import { usePanelStore } from "@/stores/panel-store";
 import { useMediaPanelStore } from "../store";
 import { AIHistoryPanel } from "./ai-history-panel";
 import { debugLogger } from "@/lib/debug-logger";
-import { getMediaStoreUtils } from "@/stores/media-store-loader";
-import { debugLog, debugError, debugWarn } from "@/lib/debug-config";
 
-// FAL API constants for testing
-const FAL_API_KEY = import.meta.env.VITE_FAL_API_KEY;
-const FAL_API_BASE = "https://fal.run";
-
-interface AIModel {
-  id: string;
-  name: string;
-  description: string;
-  price: string;
-  resolution: string;
-}
-
-const AI_MODELS: AIModel[] = [
-  {
-    id: "kling_v2",
-    name: "Kling v2.1",
-    description: "Premium model with unparalleled motion fluidity",
-    price: "0.15",
-    resolution: "1080p",
-  },
-  {
-    id: "seedance",
-    name: "Seedance v1 Lite",
-    description: "Fast and efficient text-to-video generation",
-    price: "0.18",
-    resolution: "720p",
-  },
-  {
-    id: "hailuo",
-    name: "Hailuo 02",
-    description: "Standard quality with realistic physics",
-    price: "0.27",
-    resolution: "768p",
-  },
-  {
-    id: "hailuo_pro",
-    name: "Hailuo 02 Pro",
-    description: "Premium 1080p with ultra-realistic physics",
-    price: "0.48",
-    resolution: "1080p",
-  },
-  {
-    id: "seedance_pro",
-    name: "Seedance v1 Pro",
-    description: "High quality 1080p video generation",
-    price: "0.62",
-    resolution: "1080p",
-  },
-  {
-    id: "veo3_fast",
-    name: "Veo3 Fast",
-    description: "High quality, faster generation",
-    price: "2.00",
-    resolution: "1080p",
-  },
-  {
-    id: "veo3",
-    name: "Veo3",
-    description: "Highest quality, slower generation",
-    price: "3.00",
-    resolution: "1080p",
-  },
-  {
-    id: "wan_turbo",
-    name: "WAN v2.2 Turbo",
-    description: "High-speed photorealistic video generation",
-    price: "0.10",
-    resolution: "720p",
-  },
-];
-
-interface GeneratedVideo {
-  jobId: string;
-  videoUrl: string;
-  videoPath?: string;
-  fileSize?: number;
-  duration?: number;
-  prompt: string;
-  model: string;
-}
-
-interface GeneratedVideoResult {
-  modelId: string;
-  video: GeneratedVideo;
-}
+// Import extracted hooks and types
+import { useAIGeneration } from "./use-ai-generation";
+import { useAIHistory } from "./use-ai-history";
+import { AI_MODELS, ERROR_MESSAGES, UPLOAD_CONSTANTS } from "./ai-constants";
+import type { AIActiveTab } from "./ai-types";
 
 export function AiView() {
+  // UI-only state (not related to generation logic)
   const [prompt, setPrompt] = useState("");
   const [selectedModels, setSelectedModels] = useState<string[]>([]);
-  const [isGenerating, setIsGenerating] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [jobId, setJobId] = useState<string | null>(null);
-  const [generatedVideo, setGeneratedVideo] = useState<GeneratedVideo | null>(
-    null
-  );
-  const [generatedVideos, setGeneratedVideos] = useState<
-    GeneratedVideoResult[]
-  >([]);
-  const [generationHistory, setGenerationHistory] = useState<GeneratedVideo[]>(
-    []
-  );
-
-  // AI Video Output Manager for download workflow
-  const [outputManager] = useState(
-    () => new AIVideoOutputManager("./ai-generated-videos")
-  );
-
-  // Image-to-video state
   const [selectedImage, setSelectedImage] = useState<File | null>(null);
   const [imagePreview, setImagePreview] = useState<string | null>(null);
+  const [error, setError] = useState<string | null>(null);
   const fileInputRef = useRef<HTMLInputElement>(null);
 
-  // Use global AI tab state instead of local state
-  const { aiActiveTab: activeTab, setAiActiveTab: setActiveTab } =
-    useMediaPanelStore();
+  // Use global AI tab state (CRITICAL: preserve global state integration)
+  const { aiActiveTab: activeTab, setAiActiveTab: setActiveTab } = useMediaPanelStore();
 
-  // Get project store early for debugging
+  // Get project store
   const { activeProject } = useProjectStore();
 
   // Check if current project is a fallback project
   const isFallbackProject =
     activeProject?.id?.startsWith("project-") &&
     /^project-\d{13}$/.test(activeProject?.id || "");
+
+  // Use extracted hooks
+  const generation = useAIGeneration({
+    prompt,
+    selectedModels,
+    selectedImage,
+    activeTab,
+    activeProject,
+    onProgress: (progress, message) => {
+      // Progress is handled internally by the hook
+    },
+    onError: setError,
+    onComplete: (videos) => {
+      debugLogger.log("AiView", "GENERATION_COMPLETE", {
+        videoCount: videos.length,
+        models: selectedModels
+      });
+    }
+  });
+
+  const history = useAIHistory();
+
+  // Store hooks
+  const { aiPanelWidth, aiPanelMinWidth } = usePanelStore();
+
+  // Responsive layout calculations with safe defaults
+  const safeAiPanelWidth = typeof aiPanelWidth === "number" ? aiPanelWidth : 22;
+  const safeAiPanelMinWidth =
+    typeof aiPanelMinWidth === "number" ? aiPanelMinWidth : 4;
+  const isCollapsed = safeAiPanelWidth <= safeAiPanelMinWidth + 2;
+  const isCompact = safeAiPanelWidth < 18;
+  const isExpanded = safeAiPanelWidth > 25;
 
   // Helper functions for multi-model selection
   const toggleModel = (modelId: string) => {
@@ -178,257 +102,42 @@ export function AiView() {
 
   const isModelSelected = (modelId: string) => selectedModels.includes(modelId);
 
-  // 🧪 TESTING FUNCTION: Mock generation without API calls (remove before production)
-  const handleMockGenerate = async () => {
-    if (activeTab === "text") {
-      if (!prompt.trim() || selectedModels.length === 0) return;
-    } else {
-      if (!selectedImage || selectedModels.length === 0) return;
-    }
-
-    setIsGenerating(true);
-    setError(null);
-    setJobId(null);
-    setGeneratedVideos([]);
-
-    try {
-      const mockGenerations: GeneratedVideoResult[] = [];
-
-      for (let i = 0; i < selectedModels.length; i++) {
-        const modelId = selectedModels[i];
-        const modelName = AI_MODELS.find((m) => m.id === modelId)?.name;
-
-        setStatusMessage(
-          `🧪 Mock generating with ${modelName} (${i + 1}/${selectedModels.length})`
-        );
-
-        // Simulate API delay
-        await new Promise((resolve) => setTimeout(resolve, 1500));
-
-        const mockVideo: GeneratedVideo = {
-          jobId: `mock-job-${Date.now()}-${i}`,
-          videoUrl:
-            "https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4", // More reliable test video
-          videoPath:
-            "https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4",
-          fileSize: 2_097_152,
-          duration: 15,
-          prompt: prompt.trim(),
-          model: modelId,
-        };
-
-        mockGenerations.push({ modelId, video: mockVideo });
-
-        // Mock history addition
-        addToHistory(mockVideo);
-
-        debugLogger.log("AIView", "MOCK_VIDEO_GENERATED", {
-          modelName,
-          mockJobId: mockVideo.jobId,
-          modelId,
-        });
-      }
-
-      setGeneratedVideos(mockGenerations);
-      setStatusMessage(
-        `🧪 Mock generated ${mockGenerations.length} videos successfully!`
-      );
-    } catch (error) {
-      setError(
-        "Mock generation error: " +
-          (error instanceof Error ? error.message : "Unknown error")
-      );
-      debugLogger.log("AIView", "MOCK_GENERATION_FAILED", {
-        error: error instanceof Error ? error.message : "Unknown error",
-      });
-    } finally {
-      setIsGenerating(false);
-    }
-  };
-
-  // DEBUG: Component lifecycle tracking - disabled to reduce console noise
-  // debugLogger.log("AIView", "RENDER", {
-  //   activeTab,
-  //   selectedModels,
-  //   selectedImageExists: !!selectedImage,
-  //   currentProjectId: activeProject?.id,
-  //   isFallbackProject,
-  //   currentUrl: window.location.href,
-  //   renderCount: Math.random(),
-  //   codeVersion: "2025-07-21-15:30-MULTI-SELECT-IMPLEMENTATION",
-  // });
-
-  // Temporarily disabled all window/document event monitoring for debugging
-
-  // Progress tracking
-  const [generationProgress, setGenerationProgress] = useState<number>(0);
-  const [statusMessage, setStatusMessage] = useState<string>("");
-  const [elapsedTime, setElapsedTime] = useState<number>(0);
-  const [estimatedTime, setEstimatedTime] = useState<number | undefined>();
-  const [currentModelIndex, setCurrentModelIndex] = useState<number>(0);
-  const [pollingInterval, setPollingInterval] = useState<NodeJS.Timeout | null>(
-    null
-  );
-  const [progressLogs, setProgressLogs] = useState<string[]>([]);
-  
-  // Client-side timer for elapsed time
-  const [generationStartTime, setGenerationStartTime] = useState<number | null>(null);
-
-  // Client-side elapsed time timer
-  useEffect(() => {
-    let interval: NodeJS.Timeout | null = null;
-
-    if (isGenerating && generationStartTime) {
-      interval = setInterval(() => {
-        const elapsed = Math.floor((Date.now() - generationStartTime) / 1000);
-        setElapsedTime(elapsed);
-      }, 1000); // Update every second
-    }
-
-    return () => {
-      if (interval) clearInterval(interval);
-    };
-  }, [isGenerating, generationStartTime]);
-
-  // History panel state
-  const [isHistoryPanelOpen, setIsHistoryPanelOpen] = useState<boolean>(false);
-
-  // Store hooks
-  const {
-    addMediaItem,
-    loading: mediaStoreLoading,
-    error: mediaStoreError,
-  } = useAsyncMediaStoreActions();
-  const { aiPanelWidth, aiPanelMinWidth } = usePanelStore();
-
-  // Responsive layout calculations with safe defaults
-  const safeAiPanelWidth = typeof aiPanelWidth === "number" ? aiPanelWidth : 22;
-  const safeAiPanelMinWidth =
-    typeof aiPanelMinWidth === "number" ? aiPanelMinWidth : 4;
-  const isCollapsed = safeAiPanelWidth <= safeAiPanelMinWidth + 2; // Small buffer for collapsed state
-  const isCompact = safeAiPanelWidth < 18; // Less than ~230px equivalent
-  const isExpanded = safeAiPanelWidth > 25; // Greater than ~320px equivalent
-
-  // Helper function to download video to memory
-  const downloadVideoToMemory = async (
-    videoUrl: string
-  ): Promise<Uint8Array> => {
-    debugLog("📥 Starting video download from:", videoUrl);
-
-    const response = await fetch(videoUrl);
-    if (!response.ok) {
-      throw new Error(
-        `Failed to download video: ${response.status} ${response.statusText}`
-      );
-    }
-
-    const reader = response.body?.getReader();
-    if (!reader) {
-      throw new Error("Response body is not readable");
-    }
-
-    const chunks: Uint8Array[] = [];
-    let receivedLength = 0;
-
-    try {
-      while (true) {
-        const { done, value } = await reader.read();
-        if (done) break;
-
-        chunks.push(value);
-        receivedLength += value.length;
-      }
-
-      // Combine all chunks into single Uint8Array
-      const totalData = new Uint8Array(receivedLength);
-      let position = 0;
-      for (const chunk of chunks) {
-        totalData.set(chunk, position);
-        position += chunk.length;
-      }
-
-      debugLog(`✅ Download complete: ${totalData.length} bytes total`);
-      return totalData;
-    } finally {
-      reader.releaseLock();
-    }
-  };
-
-  const maxChars = 500;
-  const remainingChars = maxChars - prompt.length;
-
-  // Load generation history from localStorage on component mount
-  useEffect(() => {
-    const savedHistory = localStorage.getItem("ai-generation-history");
-    if (savedHistory) {
-      try {
-        const parsedHistory = JSON.parse(savedHistory);
-        setGenerationHistory(parsedHistory);
-      } catch (error) {
-        debugLogger.log("AIView", "PARSE_HISTORY_ERROR", {
-          error: error instanceof Error ? error.message : "Unknown error",
-        });
-      }
-    }
-  }, []);
-
-  // Save generation history to localStorage
-  const saveGenerationHistory = (history: GeneratedVideo[]) => {
-    try {
-      localStorage.setItem("ai-generation-history", JSON.stringify(history));
-    } catch (error) {
-      debugLogger.log("AIView", "SAVE_HISTORY_ERROR", {
-        error: error instanceof Error ? error.message : "Unknown error",
-      });
-    }
-  };
-
-  // Add video to history
-  const addToHistory = (video: GeneratedVideo) => {
-    const newHistory = [video, ...generationHistory.slice(0, 9)]; // Keep only last 10
-    setGenerationHistory(newHistory);
-    saveGenerationHistory(newHistory);
-  };
-
-  // Remove video from history
-  const removeFromHistory = (jobId: string) => {
-    const newHistory = generationHistory.filter(
-      (video) => video.jobId !== jobId
-    );
-    setGenerationHistory(newHistory);
-    saveGenerationHistory(newHistory);
-  };
-
-  // Handle image selection
+  // Image handling
   const handleImageSelect = (event: React.ChangeEvent<HTMLInputElement>) => {
     const file = event.target.files?.[0];
     if (!file) return;
 
     // Validate file type
-    if (!file.type.startsWith("image/")) {
-      setError("Please select a valid image file");
+    if (!UPLOAD_CONSTANTS.ALLOWED_IMAGE_TYPES.includes(file.type as any)) {
+      setError(ERROR_MESSAGES.INVALID_FILE_TYPE);
       return;
     }
 
-    // Validate file size (10MB limit)
-    if (file.size > 10 * 1024 * 1024) {
-      setError("Image file too large (max 10MB)");
+    // Validate file size
+    if (file.size > UPLOAD_CONSTANTS.MAX_IMAGE_SIZE_BYTES) {
+      setError(ERROR_MESSAGES.FILE_TOO_LARGE);
       return;
     }
 
     setSelectedImage(file);
     setError(null);
 
-    // Create preview URL
+    // Create preview
     const reader = new FileReader();
     reader.onload = (e) => {
       setImagePreview(e.target?.result as string);
     };
     reader.readAsDataURL(file);
+
+    debugLogger.log("AiView", "IMAGE_SELECTED", {
+      fileName: file.name,
+      fileSize: file.size,
+      fileType: file.type,
+    });
   };
 
-  // Remove selected image
-  const removeSelectedImage = () => {
+  // Clear image selection
+  const clearImage = () => {
     setSelectedImage(null);
     setImagePreview(null);
     if (fileInputRef.current) {
@@ -436,369 +145,36 @@ export function AiView() {
     }
   };
 
-  // Start status polling
-  const startStatusPolling = (jobId: string) => {
-    setGenerationProgress(10);
-    setStatusMessage("Starting generation...");
-
-    const pollStatus = async () => {
-      try {
-        const status = await getGenerationStatus(jobId);
-
-        if (status.progress) {
-          setGenerationProgress(status.progress);
-        }
-
-        if (status.status === "processing") {
-          setStatusMessage(`Generating video... ${status.progress || 0}%`);
-        } else if (status.status === "completed" && status.video_url) {
-          // Clear polling
-          if (pollingInterval) {
-            clearInterval(pollingInterval);
-            setPollingInterval(null);
-          }
-
-          setGenerationProgress(100);
-          setStatusMessage("Generation complete!");
-
-          const newVideo = {
-            jobId,
-            videoUrl: status.video_url,
-            videoPath: undefined, // Not available in GenerationStatus
-            fileSize: undefined, // Not available in GenerationStatus
-            duration: undefined, // Not available in GenerationStatus
-            prompt: prompt.trim(),
-            model: selectedModels[0] || "unknown", // Use first selected model for legacy support
-          };
-
-          setGeneratedVideo(newVideo);
-          addToHistory(newVideo);
-
-          // Automatically add to media store
-          if (activeProject) {
-            try {
-              const response = await fetch(newVideo.videoUrl);
-              const blob = await response.blob();
-              const file = new File(
-                [blob],
-                `generated-video-${newVideo.jobId.substring(0, 8)}.mp4`,
-                {
-                  type: "video/mp4",
-                }
-              );
-
-              if (!addMediaItem) {
-                throw new Error("Media store not ready");
-              }
-              await addMediaItem(activeProject.id, {
-                name: `AI: ${newVideo.prompt.substring(0, 30)}...`,
-                type: "video",
-                file,
-                url: newVideo.videoUrl,
-                duration: newVideo.duration || 5,
-                width: 1920,
-                height: 1080,
-              });
-
-              debugLogger.log("AIView", "VIDEO_ADDED_TO_MEDIA_STORE", {
-                videoUrl: newVideo.videoUrl,
-                projectId: activeProject.id,
-              });
-            } catch (error) {
-              debugLogger.log("AIView", "VIDEO_ADD_TO_MEDIA_STORE_FAILED", {
-                error: error instanceof Error ? error.message : "Unknown error",
-                projectId: activeProject.id,
-              });
-            }
-          }
-
-          setIsGenerating(false);
-        } else if (status.status === "failed") {
-          // Clear polling
-          if (pollingInterval) {
-            clearInterval(pollingInterval);
-            setPollingInterval(null);
-          }
-
-          setError(status.error || "Generation failed");
-          setIsGenerating(false);
-        }
-      } catch (error) {
-        debugLogger.log("AIView", "STATUS_POLLING_ERROR", {
-          error: error instanceof Error ? error.message : "Unknown error",
-          jobId,
-        });
-        setGenerationProgress((prev) => Math.min(prev + 5, 90)); // Slowly increment until we get real status
-      }
-    };
-
-    // Poll immediately, then every 3 seconds
-    pollStatus();
-    const interval = setInterval(pollStatus, 3000);
-    setPollingInterval(interval);
-  };
-
-  // Cleanup polling on unmount
-  useEffect(() => {
-    return () => {
-      if (pollingInterval) {
-        clearInterval(pollingInterval);
-      }
-    };
-  }, [pollingInterval]);
-
-  const handleGenerate = async () => {
-    // Update validation to check for selectedModels instead of selectedModel
-    if (activeTab === "text") {
-      if (!prompt.trim() || selectedModels.length === 0) return;
-    } else {
-      if (!selectedImage || selectedModels.length === 0) return;
-    }
-
-    setIsGenerating(true);
-    setError(null);
-    setJobId(null);
-    
-    // Start the client-side timer
-    const startTime = Date.now();
-    setGenerationStartTime(startTime);
-    setElapsedTime(0);
-
-    // Reset any existing generated videos
-    setGeneratedVideos([]);
-
-    try {
-      const generations: GeneratedVideoResult[] = [];
-
-      // Sequential generation (recommended to avoid rate limits)
-      for (let i = 0; i < selectedModels.length; i++) {
-        const modelId = selectedModels[i];
-        setStatusMessage(
-          `Generating with ${AI_MODELS.find((m) => m.id === modelId)?.name} (${i + 1}/${selectedModels.length})`
-        );
-
-        let response;
-
-        setCurrentModelIndex(i);
-
-        // Create progress callback for this model
-        const progressCallback: ProgressCallback = (status) => {
-          setGenerationProgress(status.progress || 0);
-          setStatusMessage(
-            status.message ||
-              `Generating with ${AI_MODELS.find((m) => m.id === modelId)?.name}...`
-          );
-          // Note: elapsedTime is now handled by client-side timer
-          setEstimatedTime(status.estimatedTime);
-          if (status.logs) {
-            setProgressLogs(status.logs);
-          }
-        };
-
-        if (activeTab === "text") {
-          response = await generateVideo(
-            {
-              prompt: prompt.trim(),
-              model: modelId,
-              resolution: "1080p",
-              duration: 6, // Default to 6 seconds (works for all models)
-            },
-            progressCallback
-          );
-        } else {
-          response = await generateVideoFromImage({
-            image: selectedImage!,
-            model: modelId,
-            prompt: prompt.trim() || undefined,
-            resolution: "1080p",
-            duration: 6, // Default to 6 seconds (works for all models)
-          });
-        }
-
-        if (response.status === "completed" && response.video_url) {
-          const newVideo = {
-            jobId: response.job_id,
-            videoUrl: response.video_url,
-            videoPath: response.video_url,
-            fileSize: undefined,
-            duration: 6,
-            prompt: prompt.trim(),
-            model: modelId,
-          };
-
-          generations.push({ modelId, video: newVideo });
-
-          // Add each video to history as it's generated
-          addToHistory(newVideo);
-
-          // Start download workflow instead of immediate media addition
-          if (activeProject) {
-            try {
-              const modelName =
-                AI_MODELS.find((m) => m.id === modelId)?.name || modelId;
-
-              debugLog(`✅ Video generation completed for ${modelName}`);
-              debugLog(`📥 Starting download process for: ${newVideo.videoUrl}`);
-
-              // Start download tracking
-              const localPath = await outputManager.startDownload(
-                newVideo.jobId,
-                newVideo.prompt,
-                modelId
-              );
-
-              debugLog(
-                `🚀 Started download tracking for ${modelName}: ${localPath}`
-              );
-
-              // Stream download the video data
-              setStatusMessage(`Downloading ${modelName} video...`);
-              const videoData = await downloadVideoToMemory(newVideo.videoUrl);
-              
-              debugLog(`📦 Downloaded video data: ${videoData.length} bytes`);
-
-              // Complete download tracking
-              await outputManager.completeDownload(
-                newVideo.jobId,
-                newVideo.duration || 6
-              );
-
-              // Create File object from downloaded data
-              const fileName = `ai-${modelName.toLowerCase().replace(/\s+/g, "-")}-${newVideo.jobId.substring(0, 8)}.mp4`;
-              const file = await outputManager.createFileFromData(
-                videoData,
-                fileName
-              );
-
-              // Add to media panel with fully downloaded file
-              if (!addMediaItem) {
-                throw new Error("Media store not ready");
-              }
-              
-              debugLog(`💾 Adding video to media panel...`);
-              
-              // Create blob URL for immediate use
-              const blobUrl = URL.createObjectURL(file);
-              debugLog(`🔗 Created blob URL for video: ${blobUrl}`);
-              
-              // Generate thumbnail for video preview
-              let thumbnailUrl: string | undefined;
-              try {
-                debugLog(`🎬 Generating thumbnail for AI video...`);
-                const mediaUtils = await getMediaStoreUtils();
-                const videoResult = await mediaUtils.generateVideoThumbnail(file);
-                thumbnailUrl = videoResult.thumbnailUrl;
-                debugLog(`✅ Thumbnail generated successfully`);
-              } catch (error) {
-                debugWarn(`⚠️ Failed to generate thumbnail for AI video:`, error);
-              }
-              
-              await addMediaItem(activeProject.id, {
-                name: `AI (${modelName}): ${newVideo.prompt.substring(0, 20)}...`,
-                type: "video",
-                file,
-                url: blobUrl, // Add the blob URL immediately
-                thumbnailUrl, // Add the generated thumbnail
-                duration: newVideo.duration || 5,
-                width: 1920,
-                height: 1080,
-              });
-              
-              debugLog(`✅ Successfully added ${modelName} video to media panel`);
-
-              debugLogger.log(
-                "AIView",
-                "VIDEO_DOWNLOADED_AND_ADDED_TO_MEDIA_PANEL",
-                {
-                  videoUrl: newVideo.videoUrl,
-                  modelName,
-                  fileName,
-                  projectId: activeProject.id,
-                  downloadSize: videoData.length,
-                }
-              );
-            } catch (addError) {
-              outputManager.markError(
-                newVideo.jobId,
-                addError instanceof Error ? addError.message : "Unknown error"
-              );
-              debugLogger.log(
-                "AIView",
-                "VIDEO_DOWNLOAD_OR_ADD_TO_MEDIA_PANEL_FAILED",
-                {
-                  error:
-                    addError instanceof Error
-                      ? addError.message
-                      : "Unknown error",
-                  modelName: AI_MODELS.find((m) => m.id === modelId)?.name,
-                  projectId: activeProject.id,
-                }
-              );
-            }
-          }
-        }
-      }
-
-      setGeneratedVideos(generations);
-      setStatusMessage(`Generated ${generations.length} videos successfully!`);
-    } catch (error) {
-      setError(handleApiError(error));
-      debugLogger.log("AIView", "MULTI_GENERATION_FAILED", {
-        error: error instanceof Error ? error.message : "Unknown error",
-      });
-    } finally {
-      setIsGenerating(false);
-    }
-  };
-
   // Reset generation state
   const resetGenerationState = () => {
-    setGeneratedVideo(null);
-    setGeneratedVideos([]);
-    setJobId(null);
+    generation.resetGenerationState();
     setError(null);
-    setGenerationProgress(0);
-    setStatusMessage("");
-    setElapsedTime(0);
-    setEstimatedTime(undefined);
-    setCurrentModelIndex(0);
-    setProgressLogs([]);
-    setGenerationStartTime(null); // Reset timer
-    if (pollingInterval) {
-      clearInterval(pollingInterval);
-      setPollingInterval(null);
-    }
   };
 
-  const canGenerate = (() => {
-    if (isGenerating || selectedModels.length === 0) return false;
-    if (activeTab === "text") {
-      return prompt.trim().length > 0;
-    }
-    return selectedImage !== null;
-  })();
+  // Calculate cost helpers
+  const maxChars = 500;
+  const remainingChars = maxChars - prompt.length;
 
-  // Calculate total cost for selected models
   const totalCost = selectedModels.reduce((total, modelId) => {
     const model = AI_MODELS.find((m) => m.id === modelId);
     return total + (model ? parseFloat(model.price) : 0);
   }, 0);
 
   // Handle media store loading/error states
-  if (mediaStoreError) {
+  if (generation.mediaStoreError) {
     return (
       <div className="flex items-center justify-center h-full p-4">
         <div className="text-center">
           <div className="text-red-500 mb-2">Failed to load media store</div>
           <div className="text-sm text-muted-foreground">
-            {mediaStoreError.message}
+            {generation.mediaStoreError.message}
           </div>
         </div>
       </div>
     );
   }
 
-  if (mediaStoreLoading) {
+  if (generation.mediaStoreLoading) {
     return (
       <div className="flex items-center justify-center h-full">
         <div className="flex items-center space-x-2">
@@ -829,17 +205,17 @@ export function AiView() {
             </h3>
           )}
         </div>
-        {generationHistory.length > 0 && !isCollapsed && (
+        {history.hasHistory && !isCollapsed && (
           <Button
             type="button"
             size="sm"
             variant="text"
-            onClick={() => setIsHistoryPanelOpen(true)}
+            onClick={history.openHistoryPanel}
             className={`h-8 ${isCompact ? "px-1" : "px-2"}`}
           >
             <History className="size-4 mr-1" />
-            {!isCompact && `History (${generationHistory.length})`}
-            {isCompact && generationHistory.length}
+            {!isCompact && `History (${history.historyCount})`}
+            {isCompact && history.historyCount}
           </Button>
         )}
       </div>
@@ -847,307 +223,251 @@ export function AiView() {
       {/* Show collapsed state with just icon */}
       {isCollapsed ? (
         <div className="flex-1 flex flex-col items-center justify-center gap-2">
-          <BotIcon className="size-8 text-primary/50" />
-          <span className="text-[0.6rem] text-muted-foreground text-center">
-            AI
-          </span>
+          <BotIcon className="size-8 text-muted-foreground" />
+          <div className="text-xs text-muted-foreground text-center">
+            AI Video
+          </div>
         </div>
       ) : (
-        <div
-          className={`flex-1 flex flex-col ${isCompact ? "gap-2" : "gap-4"}`}
-        >
-          <Tabs
-            value={activeTab}
-            onValueChange={(value) => {
-              debugLogger.log("AIView", "TAB_CHANGE", {
-                from: activeTab,
-                to: value,
-                currentProjectId: activeProject?.id,
-              });
-              setActiveTab(value as "text" | "image");
-            }}
-          >
-            <TabsList className="grid w-full grid-cols-2 bg-transparent">
-              <TabsTrigger
-                value="text"
-                className={`flex items-center bg-transparent text-white data-[state=active]:bg-transparent data-[state=active]:text-white border-0 shadow-none hover:text-blue-400 transition-colors duration-200 ${isCompact ? "gap-1 text-xs p-2" : "gap-2"}`}
-              >
-                <TypeIcon className="size-4" />
-                {isCompact ? "Text" : "Text to Video"}
+        <div className="flex-1 overflow-y-auto space-y-4">
+          {/* Tab selector */}
+          <Tabs value={activeTab} onValueChange={(value: string) => setActiveTab(value as AIActiveTab)}>
+            <TabsList className="grid w-full grid-cols-2">
+              <TabsTrigger value="text" className="text-xs">
+                <TypeIcon className="size-3 mr-1" />
+                {!isCompact && "Text"}
               </TabsTrigger>
-              <TabsTrigger
-                value="image"
-                className={`flex items-center bg-transparent text-white data-[state=active]:bg-transparent data-[state=active]:text-white border-0 shadow-none hover:text-blue-400 transition-colors duration-200 ${isCompact ? "gap-1 text-xs p-2" : "gap-2"}`}
-              >
-                <ImageIcon className="size-4" />
-                {isCompact ? "Image" : "Image to Video"}
+              <TabsTrigger value="image" className="text-xs">
+                <ImageIcon className="size-3 mr-1" />
+                {!isCompact && "Image"}
               </TabsTrigger>
             </TabsList>
 
-            <TabsContent
-              key="text-tab-content"
-              value="text"
-              className={isCompact ? "space-y-2" : "space-y-4"}
-            >
-              <div className={isCompact ? "space-y-1" : "space-y-2"}>
-                <div style={{ height: "6px" }} />
-                <Label
-                  htmlFor="prompt"
-                  className={isCompact ? "text-xs" : ""}
-                  style={{ marginLeft: "5px" }}
-                >
-                  {isCompact ? "Prompt" : "Describe your video"}
+            <TabsContent value="text" className="space-y-4">
+              {/* Prompt input */}
+              <div className="space-y-2">
+                <Label htmlFor="prompt" className="text-xs">
+                  Prompt {!isCompact && "for Video Generation"}
                 </Label>
-                <div style={{ height: "6px" }} />
                 <Textarea
                   id="prompt"
                   placeholder={
-                    isCompact ? "Video prompt..." : "Describe your video..."
+                    isCompact
+                      ? "Describe the video..."
+                      : "Describe the video you want to generate..."
                   }
                   value={prompt}
-                  onChange={(e) => setPrompt(e.target.value.slice(0, maxChars))}
-                  className={`resize-none ${isCompact ? "min-h-[60px] text-sm" : "min-h-[80px]"}`}
+                  onChange={(e) => setPrompt(e.target.value)}
+                  className="min-h-[60px] text-xs resize-none"
+                  maxLength={maxChars}
                 />
                 <div
-                  className={`flex items-center ${isCompact ? "flex-col gap-1" : "justify-between"}`}
+                  className={`text-xs ${remainingChars < 50 ? "text-orange-500" : remainingChars < 20 ? "text-red-500" : "text-muted-foreground"} text-right`}
                 >
-                  {!isCompact && (
-                    <p
-                      className="text-[0.65rem] text-muted-foreground"
-                      style={{ marginLeft: "5px" }}
-                    >
-                      Be specific about scenes, actions, and style
-                    </p>
-                  )}
-                  <span
-                    className={`text-xs ${remainingChars < 50 ? "text-destructive" : "text-muted-foreground"}`}
-                  >
-                    {remainingChars}/{maxChars}
-                  </span>
+                  {remainingChars} characters remaining
                 </div>
               </div>
             </TabsContent>
 
-            <TabsContent
-              key="image-tab-content"
-              value="image"
-              className="space-y-4"
-            >
+            <TabsContent value="image" className="space-y-4">
+              {/* Image upload */}
               <div className="space-y-2">
-                <Label>Upload Image</Label>
-                <div style={{ height: "6px" }} />
-                {selectedImage ? (
-                  <div className="relative">
-                    <img
-                      src={imagePreview!}
-                      alt="Selected artwork"
-                      className="w-full h-40 object-cover rounded-lg border"
-                    />
-                    <Button
-                      type="button"
-                      size="sm"
-                      variant="destructive"
-                      className="absolute top-2 right-2"
-                      onClick={removeSelectedImage}
-                    >
-                      <X className="size-3" />
-                    </Button>
-                    <div className="mt-2 text-xs text-muted-foreground">
-                      {selectedImage.name} (
-                      {(selectedImage.size / 1024 / 1024).toFixed(1)} MB)
+                <Label className="text-xs">
+                  {!isCompact && "Upload "}Image{!isCompact && " for Video Generation"}
+                </Label>
+                
+                <div
+                  className={`border-2 border-dashed border-muted-foreground/25 rounded-lg p-4 text-center cursor-pointer hover:border-muted-foreground/50 transition-colors ${
+                    selectedImage ? "border-primary/50" : ""
+                  }`}
+                  onClick={() => fileInputRef.current?.click()}
+                >
+                  {selectedImage && imagePreview ? (
+                    <div className="relative">
+                      <img
+                        src={imagePreview}
+                        alt="Selected"
+                        className="max-w-full max-h-32 mx-auto rounded"
+                      />
+                      <Button
+                        type="button"
+                        size="sm"
+                        variant="destructive"
+                        onClick={(e) => {
+                          e.stopPropagation();
+                          clearImage();
+                        }}
+                        className="absolute top-1 right-1 h-6 w-6 p-0"
+                      >
+                        <X className="h-3 w-3" />
+                      </Button>
+                      <div className="mt-2 text-xs text-muted-foreground">
+                        {selectedImage.name}
+                      </div>
                     </div>
-                  </div>
-                ) : (
-                  <div
-                    className="border-2 border-dashed border-border rounded-lg p-6 text-center cursor-pointer hover:border-primary/50 transition-colors"
-                    onClick={() => fileInputRef.current?.click()}
-                  >
-                    <Upload className="size-8 mx-auto mb-2 text-muted-foreground" />
-                    <p className="text-sm text-muted-foreground mb-1">
-                      Click to upload an image
-                    </p>
-                    <p className="text-xs text-muted-foreground">
-                      JPG, PNG, WEBP (max 10MB)
-                    </p>
-                  </div>
-                )}
+                  ) : (
+                    <div className="space-y-2">
+                      <Upload className="size-8 mx-auto text-muted-foreground" />
+                      <div className="text-xs text-muted-foreground">
+                        Click to upload an image
+                      </div>
+                      <div className="text-xs text-muted-foreground/70">
+                        JPG, PNG, WebP, GIF (max 10MB)
+                      </div>
+                    </div>
+                  )}
+                </div>
+
                 <input
                   ref={fileInputRef}
                   type="file"
-                  accept="image/*"
+                  accept={UPLOAD_CONSTANTS.SUPPORTED_FORMATS.join(",")}
                   onChange={handleImageSelect}
                   className="hidden"
                 />
-              </div>
-
-              <div className="space-y-2">
-                <div style={{ height: "6px" }} />
-                <Label htmlFor="image-prompt" style={{ marginLeft: "5px" }}>
-                  Additional prompt (optional)
-                </Label>
-                <div style={{ height: "6px" }} />
-                <Textarea
-                  id="image-prompt"
-                  placeholder="Describe how the image should animate..."
-                  value={prompt}
-                  onChange={(e) => setPrompt(e.target.value.slice(0, maxChars))}
-                  className="min-h-[60px] resize-none"
-                />
-                <div className="flex justify-between items-center">
-                  <p className="text-xs text-muted-foreground">
-                    Describe motion, style, or effects
-                  </p>
-                  <span
-                    className={`text-xs ${remainingChars < 50 ? "text-destructive" : "text-muted-foreground"}`}
-                  >
-                    {remainingChars}/{maxChars}
-                  </span>
+                
+                {/* Prompt for image-to-video */}
+                <div className="space-y-2">
+                  <Label htmlFor="image-prompt" className="text-xs">
+                    {!isCompact && "Additional "}Prompt{!isCompact && " (optional)"}
+                  </Label>
+                  <Textarea
+                    id="image-prompt"
+                    placeholder={
+                      isCompact
+                        ? "Describe motion..."
+                        : "Describe how the image should move..."
+                    }
+                    value={prompt}
+                    onChange={(e) => setPrompt(e.target.value)}
+                    className="min-h-[40px] text-xs resize-none"
+                    maxLength={maxChars}
+                  />
                 </div>
               </div>
             </TabsContent>
           </Tabs>
 
-          <Button
-            onClick={handleGenerate}
-            disabled={!canGenerate}
-            className="w-full"
-            size={isCompact ? "sm" : "lg"}
-          >
-            {isGenerating ? (
-              <>
-                <Loader2 className="mr-2 size-4 animate-spin" />
-                {isCompact ? "Gen..." : "Generating..."}
-              </>
-            ) : (
-              <>
-                <BotIcon className="mr-2 size-4" />
-                {isCompact ? "Generate" : "Generate Video"}
-              </>
-            )}
-          </Button>
-
-          <div className={isCompact ? "space-y-1" : "space-y-2"}>
-            <Label
-              htmlFor="models"
-              className={`font-medium text-foreground ${isCompact ? "text-xs" : "text-sm"}`}
-            >
-              {isCompact ? "Models" : "AI Models"}
+          {/* AI Model Selection */}
+          <div className="space-y-2">
+            <Label className="text-xs">
+              {!isCompact && "Select "}AI Models{!isCompact && " (multi-select)"}
             </Label>
-            <div style={{ height: "6px" }} />
-            <div
-              className={`grid grid-cols-1 gap-2 overflow-y-auto ${isCompact ? "max-h-32" : "max-h-48"}`}
-            >
+            <div className="space-y-1">
               {AI_MODELS.map((model) => (
-                <Button
+                <div
                   key={model.id}
-                  type="button"
-                  variant="outline"
-                  size="sm"
-                  onClick={() => toggleModel(model.id)}
-                  className={`
-                  flex items-center justify-between h-auto text-left font-mono
-                  transition-all duration-200 border-border/50 
-                  ${isCompact ? "p-2" : "p-3"}
-                  ${
+                  className={`flex items-center justify-between p-2 rounded-md border cursor-pointer transition-colors ${
                     isModelSelected(model.id)
-                      ? "bg-transparent border-blue-500/50"
-                      : "bg-transparent hover:bg-accent/50 hover:border-border"
-                  }
-                `}
+                      ? "border-primary bg-primary/5"
+                      : "border-border hover:border-muted-foreground/50"
+                  }`}
+                  onClick={() => toggleModel(model.id)}
                 >
-                  <div
-                    className={`flex items-center ${isCompact ? "gap-2" : "gap-3"}`}
-                  >
+                  <div className="flex items-center space-x-2">
                     <div
-                      className={`
-                    w-4 h-4 rounded border flex items-center justify-center
-                    ${
-                      isModelSelected(model.id)
-                        ? "bg-blue-500 border-transparent"
-                        : "border-border bg-transparent"
-                    }
-                  `}
+                      className={`w-4 h-4 rounded border-2 flex items-center justify-center ${
+                        isModelSelected(model.id)
+                          ? "border-primary bg-primary"
+                          : "border-muted-foreground/30"
+                      }`}
                     >
                       {isModelSelected(model.id) && (
-                        <Check className="w-3 h-3 text-white" />
+                        <Check className="w-2.5 h-2.5 text-primary-foreground" />
                       )}
                     </div>
-                    <div className="flex flex-col">
-                      <span
-                        className={`font-medium ${isCompact ? "text-xs" : "text-sm"} ${isModelSelected(model.id) ? "text-[#05c7c7]" : "text-foreground"}`}
-                      >
-                        {isCompact ? model.name.split(" ")[0] : model.name}
-                      </span>
+                    <div>
+                      <div className="text-xs font-medium">{model.name}</div>
                       {!isCompact && (
-                        <span
-                          className={`text-xs ${isModelSelected(model.id) ? "text-[#05c7c7]" : "text-muted-foreground"}`}
-                        >
+                        <div className="text-xs text-muted-foreground">
                           {model.description}
-                        </span>
+                        </div>
                       )}
                     </div>
                   </div>
-                  {!isCompact && (
-                    <span
-                      className={`text-xs font-normal ml-6 ${isModelSelected(model.id) ? "text-[#05c7c7]" : "text-muted-foreground"}`}
-                    >
-                      USD {model.price} • {model.resolution}
-                    </span>
-                  )}
-                  {isCompact && (
-                    <span
-                      className={`text-xs font-normal ${isModelSelected(model.id) ? "text-[#05c7c7]" : "text-muted-foreground"}`}
-                    >
-                      ${model.price}
-                    </span>
-                  )}
-                </Button>
+                  <div className="text-xs text-right">
+                    <div>${model.price}</div>
+                    {!isCompact && (
+                      <div className="text-muted-foreground">{model.resolution}</div>
+                    )}
+                  </div>
+                </div>
               ))}
             </div>
-
-            <div className="flex gap-2 pt-2">
-              <Button
-                type="button"
-                size="sm"
-                variant="outline"
-                onClick={() => setSelectedModels(AI_MODELS.map((m) => m.id))}
-                className="text-xs flex-1 !bg-transparent"
-              >
-                Select All
-              </Button>
-              <Button
-                type="button"
-                size="sm"
-                variant="outline"
-                onClick={() => setSelectedModels([])}
-                className="text-xs flex-1 !bg-transparent"
-              >
-                Clear All
-              </Button>
-            </div>
+            
+            {/* Cost display */}
+            {selectedModels.length > 0 && (
+              <div className="text-xs text-muted-foreground text-right">
+                Total estimated cost: <span className="font-medium">${totalCost.toFixed(2)}</span>
+              </div>
+            )}
           </div>
 
-          {selectedModels.length > 0 && (
-            <div className="bg-panel-accent rounded-lg p-3">
-              <div className="flex items-center gap-2 mb-2">
-                <BotIcon className="size-4 text-primary" />
-                <span className="text-sm font-medium">
-                  {selectedModels.length} Model
-                  {selectedModels.length > 1 ? "s" : ""} Selected
-                </span>
+          {/* Error display */}
+          {error && (
+            <div className="p-3 bg-destructive/10 border border-destructive/20 rounded-md">
+              <div className="text-xs text-destructive">{error}</div>
+            </div>
+          )}
+
+          {/* Progress display */}
+          {generation.isGenerating && (
+            <div className="space-y-3 p-3 bg-muted/50 rounded-md">
+              <div className="flex items-center justify-between">
+                <div className="text-xs font-medium">Generating...</div>
+                <div className="text-xs text-muted-foreground">
+                  {Math.round(generation.generationProgress)}%
+                </div>
               </div>
-              <div className="space-y-1">
-                {selectedModels.map((modelId) => {
-                  const model = AI_MODELS.find((m) => m.id === modelId);
+              <div className="w-full bg-muted-foreground/20 rounded-full h-2">
+                <div
+                  className="bg-primary h-2 rounded-full transition-all duration-300"
+                  style={{ width: `${generation.generationProgress}%` }}
+                />
+              </div>
+              <div className="text-xs text-muted-foreground">
+                {generation.statusMessage}
+              </div>
+              {generation.elapsedTime > 0 && (
+                <div className="text-xs text-muted-foreground">
+                  Elapsed: {Math.floor(generation.elapsedTime / 60)}:
+                  {(generation.elapsedTime % 60).toString().padStart(2, "0")}
+                </div>
+              )}
+            </div>
+          )}
+
+          {/* Generated videos results */}
+          {generation.hasResults && (
+            <div className="space-y-2">
+              <Label className="text-xs">Generated Videos</Label>
+              <div className="space-y-2">
+                {generation.generatedVideos.map((result, index) => {
+                  const model = AI_MODELS.find((m) => m.id === result.modelId);
                   return (
-                    <div
-                      key={modelId}
-                      className="flex justify-between items-center text-xs"
-                    >
-                      <span className="text-foreground">{model?.name}</span>
-                      <span className="text-muted-foreground">
-                        USD {model?.price}
-                      </span>
+                    <div key={index} className="flex items-center justify-between p-2 bg-muted/30 rounded-md">
+                      <div className="flex items-center space-x-2">
+                        <Play className="size-4 text-primary" />
+                        <div>
+                          <div className="text-xs font-medium">{model?.name}</div>
+                          <div className="text-xs text-muted-foreground">
+                            {result.video.prompt.substring(0, 30)}...
+                          </div>
+                        </div>
+                      </div>
+                      <Button
+                        size="sm"
+                        variant="outline"
+                        onClick={() => {
+                          const a = document.createElement("a");
+                          a.href = result.video.videoUrl;
+                          a.download = `ai-video-${result.video.jobId}.mp4`;
+                          a.click();
+                        }}
+                        className="h-6 px-2"
+                      >
+                        <Download className="size-3" />
+                      </Button>
                     </div>
                   );
                 })}
@@ -1155,298 +475,68 @@ export function AiView() {
             </div>
           )}
 
-          <div className="mt-auto pt-4 space-y-2">
+          {/* Action buttons */}
+          <div className="space-y-2 pt-2">
             <Button
-              onClick={handleMockGenerate}
-              disabled={!canGenerate}
+              onClick={generation.handleGenerate}
+              disabled={!generation.canGenerate}
               className="w-full"
-              size="lg"
-              variant="outline"
+              size={isCompact ? "sm" : "lg"}
             >
-              {isGenerating ? (
+              {generation.isGenerating ? (
                 <>
-                  <Loader2 className="mr-2 size-4 animate-spin" />
-                  Generating...
+                  <Loader2 className="size-4 mr-2 animate-spin" />
+                  {isCompact ? "Generating..." : "Generating Video..."}
                 </>
               ) : (
-                <>Generate Video</>
+                <>
+                  <BotIcon className="size-4 mr-2" />
+                  {isCompact ? "Generate" : "Generate Video"}
+                </>
               )}
             </Button>
+
+            {/* Mock generation for testing */}
+            {process.env.NODE_ENV === "development" && (
+              <Button
+                onClick={generation.handleMockGenerate}
+                disabled={!generation.canGenerate}
+                className="w-full"
+                size="lg"
+                variant="outline"
+              >
+                🧪 Mock Generate (Dev)
+              </Button>
+            )}
+
+            {/* Reset button */}
+            {(generation.hasResults || error) && (
+              <Button
+                onClick={resetGenerationState}
+                variant="outline"
+                size="sm"
+                className="w-full"
+              >
+                <X className="size-3 mr-1" />
+                Reset
+              </Button>
+            )}
           </div>
-
-          {isGenerating && (
-            <div className="mt-4 p-4 bg-blue-500/10 border border-blue-500/20 rounded-lg space-y-3">
-              <div className="flex items-center justify-between">
-                <div className="flex items-center gap-2">
-                  <Loader2 className="size-4 animate-spin text-blue-500" />
-                  <span className="text-sm font-medium text-blue-700">
-                    Model {currentModelIndex + 1} of {selectedModels.length}
-                  </span>
-                </div>
-                <div className="text-xs text-blue-600">
-                  {Math.floor(elapsedTime / 60)}:
-                  {String(elapsedTime % 60).padStart(2, "0")}
-                </div>
-              </div>
-
-              <div className="space-y-2">
-                <div className="flex items-center justify-between text-xs">
-                  <span className="text-blue-600">{statusMessage}</span>
-                  <span className="text-blue-600">{generationProgress}%</span>
-                </div>
-                <div className="w-full bg-blue-100 rounded-full h-2">
-                  <div
-                    className="bg-blue-500 h-2 rounded-full transition-all duration-300 ease-out"
-                    style={{ width: `${generationProgress}%` }}
-                  />
-                </div>
-              </div>
-
-              <div className="flex justify-between text-xs text-blue-600">
-                <span>
-                  Elapsed: {Math.floor(elapsedTime / 60)}:
-                  {String(elapsedTime % 60).padStart(2, "0")}
-                </span>
-                {estimatedTime && (
-                  <span>
-                    Est. remaining: {Math.floor(estimatedTime / 60)}:
-                    {String(estimatedTime % 60).padStart(2, "0")}
-                  </span>
-                )}
-              </div>
-
-              <div className="text-xs text-blue-600">
-                Generating with:{" "}
-                <span className="font-medium">
-                  {
-                    AI_MODELS.find(
-                      (m) => m.id === selectedModels[currentModelIndex]
-                    )?.name
-                  }
-                </span>
-              </div>
-
-              {progressLogs.length > 0 && (
-                <details className="text-xs">
-                  <summary className="text-blue-600 cursor-pointer hover:text-blue-700">
-                    View detailed logs ({progressLogs.length} entries)
-                  </summary>
-                  <div className="mt-2 p-2 bg-blue-50 rounded border max-h-24 overflow-y-auto">
-                    {progressLogs.map((log, index) => (
-                      <div
-                        key={index}
-                        className="text-blue-700 font-mono text-[10px] mb-1"
-                      >
-                        {log}
-                      </div>
-                    ))}
-                  </div>
-                </details>
-              )}
-            </div>
-          )}
-
-          {error && (
-            <div className="mt-2 p-2 bg-destructive/10 border border-destructive/20 rounded text-sm text-destructive">
-              {error}
-            </div>
-          )}
-
-          {isGenerating && jobId && (
-            <div className="mt-2 space-y-2">
-              <div className="p-3 bg-blue-500/10 border border-blue-500/20 rounded">
-                <div className="flex items-center gap-2 mb-2">
-                  <Loader2 className="size-4 animate-spin text-blue-600" />
-                  <span className="text-sm font-medium text-blue-700">
-                    {statusMessage || "Processing..."}
-                  </span>
-                </div>
-                <div className="w-full bg-gray-200 rounded-full h-2">
-                  <div
-                    className="bg-blue-600 h-2 rounded-full transition-all duration-300"
-                    style={{ width: `${generationProgress}%` }}
-                  />
-                </div>
-                <div className="flex justify-between items-center mt-1">
-                  <span className="text-xs text-blue-600">
-                    Job ID: {jobId.substring(0, 8)}...
-                  </span>
-                  <span className="text-xs text-blue-600">
-                    {generationProgress}%
-                  </span>
-                </div>
-              </div>
-            </div>
-          )}
-
-          {jobId && !isGenerating && !generatedVideo && !error && (
-            <div className="mt-2 p-2 bg-green-500/10 border border-green-500/20 rounded text-sm text-green-700">
-              ✅ Generation completed! Processing video...
-            </div>
-          )}
-
-          {selectedModels.length > 0 && !generatedVideos.length && (
-            <div className="mt-2 text-center">
-              <span className="text-xs text-muted-foreground">
-                Total Cost: USD {totalCost.toFixed(2)} • {selectedModels.length}{" "}
-                model{selectedModels.length > 1 ? "s" : ""}
-              </span>
-            </div>
-          )}
-
-          {!canGenerate && !isGenerating && generatedVideos.length === 0 && (
-            <div className="mt-2 text-center">
-              <span className="text-xs text-muted-foreground">
-                {selectedModels.length === 0
-                  ? "Select at least one AI model"
-                  : activeTab === "text"
-                    ? "Enter a video description"
-                    : "Upload an image"}
-              </span>
-            </div>
-          )}
-
-          {generatedVideos.length > 0 && (
-            <div className="mt-4 space-y-3">
-              <div className="flex items-center gap-2 mb-3">
-                <Play className="size-4 text-green-600" />
-                <span className="text-sm font-medium text-green-700">
-                  {generatedVideos.length} Video
-                  {generatedVideos.length > 1 ? "s" : ""} Generated
-                  Successfully!
-                </span>
-              </div>
-
-              <div className="text-sm text-green-700 mb-3">
-                ✅ All videos automatically added to Media panel
-              </div>
-
-              <div className="space-y-2">
-                {generatedVideos.map(({ modelId, video }) => {
-                  const model = AI_MODELS.find((m) => m.id === modelId);
-                  return (
-                    <div
-                      key={modelId}
-                      className="bg-green-500/5 border border-green-500/20 rounded-lg p-3"
-                    >
-                      <div className="flex justify-between items-center mb-2">
-                        <span className="font-medium text-sm text-green-700">
-                          {model?.name}
-                        </span>
-                        <span className="text-xs text-muted-foreground">
-                          USD {model?.price} • {model?.resolution}
-                        </span>
-                      </div>
-
-                      <div className="flex gap-2">
-                        <Button
-                          type="button"
-                          size="sm"
-                          variant="outline"
-                          onClick={() => {
-                            const link = document.createElement("a");
-                            link.href = video.videoUrl;
-                            link.download = `ai-${model?.name.toLowerCase().replace(/\s+/g, "-")}-${video.jobId.substring(0, 8)}.mp4`;
-                            link.click();
-                          }}
-                          className="flex-1"
-                        >
-                          <Download className="mr-1 size-3" />
-                          Download
-                        </Button>
-                      </div>
-                    </div>
-                  );
-                })}
-              </div>
-
-              <div className="flex gap-2 pt-2">
-                <Button
-                  type="button"
-                  size="sm"
-                  variant="outline"
-                  onClick={() => {
-                    // Download all videos
-                    generatedVideos.forEach(({ modelId, video }) => {
-                      const model = AI_MODELS.find((m) => m.id === modelId);
-                      const link = document.createElement("a");
-                      link.href = video.videoUrl;
-                      link.download = `ai-${model?.name.toLowerCase().replace(/\s+/g, "-")}-${video.jobId.substring(0, 8)}.mp4`;
-                      link.click();
-                    });
-                  }}
-                  className="flex-1"
-                >
-                  <Download className="mr-1 size-3" />
-                  Download All ({generatedVideos.length})
-                </Button>
-                <Button
-                  type="button"
-                  size="sm"
-                  variant="outline"
-                  onClick={resetGenerationState}
-                  className="flex-1"
-                >
-                  Generate Again
-                </Button>
-              </div>
-            </div>
-          )}
-
-          {generatedVideo && generatedVideos.length === 0 && (
-            <div className="mt-4 p-4 bg-green-500/10 border border-green-500/20 rounded-lg">
-              <div className="flex items-center gap-2 mb-3">
-                <Play className="size-4 text-green-600" />
-                <span className="text-sm font-medium text-green-700">
-                  Video Generated Successfully!
-                </span>
-              </div>
-
-              <div className="text-sm text-green-700 mb-3">
-                ✅ Video automatically added to Media panel
-              </div>
-
-              <div className="flex gap-2">
-                <Button
-                  type="button"
-                  size="sm"
-                  variant="outline"
-                  onClick={() => {
-                    const link = document.createElement("a");
-                    link.href = generatedVideo.videoUrl;
-                    link.download = `generated-video-${generatedVideo.jobId.substring(0, 8)}.mp4`;
-                    link.click();
-                  }}
-                >
-                  <Download className="mr-1 size-3" />
-                  Download
-                </Button>
-                <Button
-                  type="button"
-                  size="sm"
-                  variant="outline"
-                  className="flex-1"
-                  onClick={resetGenerationState}
-                >
-                  Generate Another Video
-                </Button>
-              </div>
-            </div>
-          )}
-
-          <AIHistoryPanel
-            isOpen={isHistoryPanelOpen}
-            onClose={() => setIsHistoryPanelOpen(false)}
-            generationHistory={generationHistory}
-            onSelectVideo={(video) => {
-              setGeneratedVideo(video);
-              setIsHistoryPanelOpen(false);
-            }}
-            onRemoveFromHistory={removeFromHistory}
-            aiModels={AI_MODELS}
-          />
         </div>
       )}
+
+      {/* History Panel */}
+      <AIHistoryPanel
+        isOpen={history.isHistoryPanelOpen}
+        onClose={history.closeHistoryPanel}
+        generationHistory={history.generationHistory}
+        onSelectVideo={(video) => {
+          generation.setGeneratedVideo(video);
+          history.closeHistoryPanel();
+        }}
+        onRemoveFromHistory={history.removeFromHistory}
+        aiModels={AI_MODELS}
+      />
     </div>
   );
 }
